# docs/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Bank_REST API","description":"API for bank cards and bank users","version":"1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"User","description":"User management endpoints"},{"name":"Card","description":"Card management endpoints"},{"name":"Auth","description":"Authentication endpoints"}],"paths":{"/user/update":{"put":{"tags":["User"],"summary":"Update user","description":"Updates user data. Only for ADMIN role.","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"404":{"description":"User not found"},"200":{"description":"User updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/card/update":{"put":{"tags":["Card"],"summary":"Update card","description":"Updates card data. Only for ADMIN role.","operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardRequest"}}},"required":true},"responses":{"200":{"description":"Card updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"404":{"description":"Card not found"}}}},"/user/save":{"post":{"tags":["User"],"summary":"Create new user","description":"Creates a new user. Only for ADMIN role.","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveUserRequest"}}},"required":true},"responses":{"400":{"description":"Invalid input"},"200":{"description":"User created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/addRole":{"post":{"tags":["User"],"summary":"Add new role","description":"Adds a new role to the system. Only for ADMIN role.","operationId":"addRole","parameters":[{"name":"roleName","in":"query","description":"Role name","required":true,"schema":{"type":"string"},"example":"MANAGER"}],"responses":{"200":{"description":"Role added"},"400":{"description":"Invalid input"}}}},"/card/transfer":{"post":{"tags":["Card"],"summary":"Transfer between cards","description":"Transfers money between cards. Only for authenticated users.","operationId":"transferBetweenCards","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferBetweenCardsRequest"}}},"required":true},"responses":{"400":{"description":"Invalid input"},"200":{"description":"Transfer successful"}}}},"/card/save":{"post":{"tags":["Card"],"summary":"Create new card","description":"Creates a new card. Only for ADMIN role.","operationId":"save_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveCardRequest"}}},"required":true},"responses":{"200":{"description":"Card created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"400":{"description":"Invalid input"}}}},"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","description":"Registers a new user.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveUserRequest"}}}},"responses":{"400":{"description":"Invalid input"},"200":{"description":"User registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Authenticate user and return JWT token.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"Authentication successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}},"401":{"description":"Invalid credentials"}}}},"/card/getBalance":{"patch":{"tags":["Card"],"summary":"Get card balance","description":"Returns card balance. Only for authenticated users.","operationId":"getBalance","parameters":[{"name":"id","in":"query","description":"Card ID","required":true,"schema":{"type":"integer","format":"int64"},"example":100}],"responses":{"404":{"description":"Card not found"},"200":{"description":"Balance returned"}}}},"/card/blockCard":{"patch":{"tags":["Card"],"summary":"Block card","description":"Blocks a card. Only for authenticated users.","operationId":"blockCard","parameters":[{"name":"id","in":"query","description":"Card ID","required":true,"schema":{"type":"integer","format":"int64"},"example":100}],"responses":{"200":{"description":"Card blocked"},"404":{"description":"Card not found"}}}},"/user/getByUserName":{"get":{"tags":["User"],"summary":"Get user by username","description":"Returns user by username. Only for ADMIN role.","operationId":"getByUsername","parameters":[{"name":"username","in":"query","description":"Username","required":true,"schema":{"type":"string"},"example":"john_doe"}],"responses":{"404":{"description":"User not found"},"200":{"description":"User found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/getById":{"get":{"tags":["User"],"summary":"Get user by ID","description":"Returns user by ID. Only for ADMIN role.","operationId":"getById","parameters":[{"name":"id","in":"query","description":"User ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"User not found"},"200":{"description":"User found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/user/getAll":{"get":{"tags":["User"],"summary":"Get all users","description":"Returns all users with optional filtering. Only for ADMIN role.","operationId":"getAll","parameters":[{"name":"filter","in":"query","description":"User filter","required":true,"schema":{"$ref":"#/components/schemas/UserFilter"}},{"name":"offset","in":"query","description":"Page offset","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"limit","in":"query","description":"Page limit","required":false,"schema":{"type":"integer","format":"int32","default":20},"example":20}],"responses":{"200":{"description":"List of users","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/card/getById":{"get":{"tags":["Card"],"summary":"Get card by ID","description":"Returns card by ID. Only for ADMIN role.","operationId":"getById_1","parameters":[{"name":"id","in":"query","description":"Card ID","required":true,"schema":{"type":"integer","format":"int64"},"example":100}],"responses":{"200":{"description":"Card found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}},"404":{"description":"Card not found"}}}},"/card/getAll":{"get":{"tags":["Card"],"summary":"Get all cards","description":"Returns all cards with optional filtering. Only for ADMIN role.","operationId":"getAll_1","parameters":[{"name":"filter","in":"query","description":"Card filter","required":true,"schema":{"$ref":"#/components/schemas/CardFilter"}},{"name":"offset","in":"query","description":"Page offset","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"limit","in":"query","description":"Page limit","required":false,"schema":{"type":"integer","format":"int32","default":20},"example":20}],"responses":{"200":{"description":"List of cards","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/card/getAllHolderCards":{"get":{"tags":["Card"],"summary":"Get all cards of current user","description":"Returns all cards of the authenticated user.","operationId":"getAllHolderCards","parameters":[{"name":"filter","in":"query","description":"Card filter","required":true,"schema":{"$ref":"#/components/schemas/CardFilter"}},{"name":"offset","in":"query","description":"Page offset","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"limit","in":"query","description":"Page limit","required":false,"schema":{"type":"integer","format":"int32","default":20},"example":20}],"responses":{"200":{"description":"List of cards","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/user/deleteById":{"delete":{"tags":["User"],"summary":"Delete user by ID","description":"Deletes a user by ID. Only for ADMIN role.","operationId":"deleteById","parameters":[{"name":"id","in":"query","description":"User ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"User deleted","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}},"404":{"description":"User not found"}}}},"/user/deleteAll":{"delete":{"tags":["User"],"summary":"Delete all users","description":"Deletes all users. Only for ADMIN role.","operationId":"deleteAll","responses":{"200":{"description":"All users deleted"}}}},"/card/deleteById":{"delete":{"tags":["Card"],"summary":"Delete card by ID","description":"Deletes a card by ID. Only for ADMIN role.","operationId":"deleteById_1","parameters":[{"name":"id","in":"query","description":"Card ID","required":true,"schema":{"type":"integer","format":"int64"},"example":100}],"responses":{"200":{"description":"Card deleted"},"404":{"description":"Card not found"}}}},"/card/deleteAll":{"delete":{"tags":["Card"],"summary":"Delete all cards","description":"Deletes all cards. Only for ADMIN role.","operationId":"deleteAll_1","responses":{"200":{"description":"All cards deleted"}}}}},"components":{"schemas":{"UpdateUserRequest":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","description":"User ID","format":"int64","example":1},"password":{"type":"string","description":"New password","example":"newpassword123"},"roles":{"uniqueItems":true,"type":"array","description":"New roles","example":"[USER, ADMIN]","items":{"type":"string","description":"New roles","example":"[USER, ADMIN]"}}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","description":"User ID","format":"int64","example":1},"username":{"type":"string","description":"Username","example":"john_doe"},"password":{"type":"string","description":"User password","example":"password123"},"roles":{"uniqueItems":true,"type":"array","description":"User roles","example":["ADMIN","USER"],"items":{"type":"string","description":"User roles","example":"[\"ADMIN\",\"USER\"]"}}}},"UpdateCardRequest":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","description":"Card ID","format":"int64","example":100},"newHolderId":{"type":"integer","description":"New holder ID","format":"int64","example":2},"newExpirationDate":{"type":"string","description":"New expiration date","format":"date","example":"2026-01-01"},"newStatus":{"type":"string","description":"New card status","example":"Blocked","enum":["Active","Blocked","Expired"]},"newBalance":{"type":"number","description":"New card balance","example":500.0}}},"CardDto":{"type":"object","properties":{"id":{"type":"integer","description":"Card ID","format":"int64","example":100},"number":{"type":"string","description":"Card number","example":"1234-5678-9012-3456"},"holderName":{"type":"string","description":"Card holder name","example":"John Doe"},"expirationDate":{"type":"string","description":"Expiration date","format":"date","example":"2025-12-31"},"status":{"type":"string","description":"Card status","example":"Active","enum":["Active","Blocked","Expired"]},"balance":{"type":"number","description":"Card balance","example":1000.0}}},"SaveUserRequest":{"required":["password","roles","userName"],"type":"object","properties":{"userName":{"type":"string","description":"Username","example":"admin"},"password":{"type":"string","description":"Password","example":"password123"},"roles":{"uniqueItems":true,"type":"array","description":"Roles","example":"[USER, ADMIN]","items":{"type":"string","description":"Roles","example":"[USER, ADMIN]"}}}},"TransferBetweenCardsRequest":{"required":["amount","fromCardId","toCardId"],"type":"object","properties":{"fromCardId":{"type":"integer","description":"Source card ID","format":"int64","example":100},"toCardId":{"type":"integer","description":"Destination card ID","format":"int64","example":101},"amount":{"type":"number","description":"Transfer amount","example":250.0}}},"SaveCardRequest":{"required":["holderId"],"type":"object","properties":{"holderId":{"type":"integer","description":"Holder ID","format":"int64","example":1}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}},"UserFilter":{"type":"object","properties":{"username":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"CardFilter":{"type":"object","properties":{"number":{"type":"string"},"expirationDate":{"type":"string","format":"date"},"status":{"type":"string","enum":["Active","Blocked","Expired"]},"balance":{"type":"number"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}